fastapi/requirements.txt
fastapi~=0.95.0
uvicorn~=0.22.0
sqlalchemy~=2.0.0
alembic~=1.10.0
pydantic~=1.10.0
email-validator~=2.0.0
pytest~=7.3.0
httpx~=0.24.0
python-dotenv~=1.0.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
psycopg~=3.2.3
